# cmd/docs/v2/swagger.yaml
basePath: /
definitions:
...
  dtos.UpdateImageDTO:
    properties:
      image_url:
        type: string
    type: object
...
  dtos.UpdateUserDTO:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      message:
        example: message
        type: string
      status_code:
        example: 888
        type: integer
    type: object
host: localhost:5000/api/v2
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API documentation for LeaseEase.
  title: LeaseEase API
  version: "2.0"
paths:
...
  /user/check:
    post:
      consumes:
      - application/json
      description: Validates the auth token from cookies and returns user information
        if authenticated
      produces:
      - application/json
      responses:
        "200":
          description: User is authenticated with claims data
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized - No token provided or invalid token
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - CookieAuth: []
      summary: Verify user authentication
      tags:
      - users
  /user/get/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves user details for the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to retrieve user
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - CookieAuth: []
      summary: Get user information
      tags:
      - users
  /user/image:
    put:
      consumes:
      - application/json
      description: Updates the profile image for the authenticated user
      parameters:
      - description: Image update information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateImageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Image updated successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to update image
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - CookieAuth: []
      summary: Update user profile image
      tags:
      - users
  /user/user:
    put:
      consumes:
      - application/json
      description: Updates user details for the authenticated user
      parameters:
      - description: User update information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - CookieAuth: []
      summary: Update user information
      tags:
      - users
swagger: "2.0"